+++ egui_commonmark_macros/src/generator.rs
@@ -158,6 +158,12 @@ pub struct StyledImage {
     pub alt_text: Vec<StyledText>,
 }
 
+#[derive(Default)]
+struct DefinitionList {
+    is_first_item: bool,
+    is_def_list_def: bool,
+}
+
 pub(crate) struct CommonMarkViewerInternal {
     source_id: Id,
     curr_table: usize,
@@ -168,6 +174,7 @@ pub(crate) struct CommonMarkViewerInternal {
     line: Newline,
     fenced_code_block: Option<FencedCodeBlock>,
     is_list_item: bool,
+    def_list: DefinitionList,
     is_table: bool,
     is_blockquote: bool,
 
@@ -188,6 +195,7 @@ impl CommonMarkViewerInternal {
             image: None,
             line: Newline::default(),
             is_list_item: false,
+            def_list: Default::default(),
             fenced_code_block: None,
             is_table: false,
             is_blockquote: false,
@@ -264,8 +272,69 @@ impl CommonMarkViewerInternal {
         let mut stream = self.event(event, cache, options);
 
         stream.extend(self.item_list_wrapping(events, cache, options));
+        stream.extend(self.def_list_def_wrapping(events, cache, options));
         stream.extend(self.table(events, cache, options));
         stream.extend(self.blockquote(events, cache, options));
+        stream
+    }
+
+    fn def_list_def_wrapping<'e>(
+        &mut self,
+        events: &mut Peekable<impl Iterator<Item = EventIteratorItem<'e>>>,
+        cache: &Expr,
+        options: &CommonMarkOptions,
+    ) -> TokenStream {
+        let mut stream = TokenStream::new();
+        if self.def_list.is_def_list_def {
+            self.def_list.is_def_list_def = false;
+
+            let item_events = delayed_events(events, |tag| {
+                matches!(tag, pulldown_cmark::TagEnd::DefinitionListDefinition)
+            });
+
+            let mut events_iter = item_events.into_iter().enumerate().peekable();
+
+            let mut inner = TokenStream::new();
+
+            stream.extend(self.line.try_insert_start());
+
+            // Proccess a single event separately so that we do not insert spaces where we do not
+            // want them
+            self.line.should_start_newline = false;
+            if let Some((_, (e, _))) = events_iter.next() {
+                inner.extend(self.process_event(&mut events_iter, e, cache, options));
+            }
+
+            self.line.should_start_newline = true;
+            self.line.should_end_newline = false;
+            while let Some((_, (e, _))) = events_iter.next() {
+                inner.extend(self.process_event(&mut events_iter, e, cache, options));
+            }
+            self.line.should_end_newline = true;
+
+            let spaces = " ".repeat(options.indentation_spaces);
+            stream.extend(quote!(ui.label(#spaces);));
+
+            // Required to ensure that the content is aligned with the identation
+            stream.extend(quote!(ui.horizontal_wrapped(|ui| {
+                    #inner
+            });));
+
+            // Only end the definition items line if it is not the last element in the list
+            if !matches!(
+                events.peek(),
+                Some((
+                    _,
+                    (
+                        pulldown_cmark::Event::End(pulldown_cmark::TagEnd::DefinitionList),
+                        _
+                    )
+                ))
+            ) {
+                stream.extend(self.line.try_insert_end());
+            }
+        }
+
         stream
     }
 
@@ -306,8 +375,9 @@ impl CommonMarkViewerInternal {
     ) -> TokenStream {
         let mut stream = TokenStream::new();
         if self.is_blockquote {
-            let mut collected_events =
-                delayed_events(events, pulldown_cmark::TagEnd::BlockQuote(None));
+            let mut collected_events = delayed_events(events, |tag| {
+                matches!(tag, pulldown_cmark::TagEnd::BlockQuote(_))
+            });
             stream.extend(self.line.try_insert_start());
 
             self.line.should_start_newline = true;
@@ -613,9 +683,23 @@ impl CommonMarkViewerInternal {
             pulldown_cmark::Tag::HtmlBlock | pulldown_cmark::Tag::MetadataBlock(_) => {
                 TokenStream::new()
             }
-            pulldown_cmark::Tag::DefinitionList => TokenStream::new(),
-            pulldown_cmark::Tag::DefinitionListTitle => TokenStream::new(),
-            pulldown_cmark::Tag::DefinitionListDefinition => TokenStream::new(),
+            pulldown_cmark::Tag::DefinitionList => {
+                let s = self.line.try_insert_start();
+                self.def_list.is_first_item = true;
+                s
+            }
+            pulldown_cmark::Tag::DefinitionListTitle => {
+                if !self.def_list.is_first_item {
+                    self.line.try_insert_start()
+                } else {
+                    self.def_list.is_first_item = false;
+                    TokenStream::new()
+                }
+            }
+            pulldown_cmark::Tag::DefinitionListDefinition => {
+                self.def_list.is_def_list_def = true;
+                TokenStream::new()
+            }
         }
     }
 
@@ -716,7 +800,7 @@ impl CommonMarkViewerInternal {
             pulldown_cmark::TagEnd::HtmlBlock | pulldown_cmark::TagEnd::MetadataBlock(_) => {
                 TokenStream::new()
             }
-            pulldown_cmark::TagEnd::DefinitionList => TokenStream::new(),
+            pulldown_cmark::TagEnd::DefinitionList => self.line.try_insert_end(),
             pulldown_cmark::TagEnd::DefinitionListTitle => TokenStream::new(),
             pulldown_cmark::TagEnd::DefinitionListDefinition => TokenStream::new(),
         }